+++ lib/oxidized/output/git.rb
@@ -167,7 +167,7 @@ module Oxidized
 
       begin
         repo = Rugged::Repository.new repo
-        update_repo repo, file, data, @msg, @user, @email
+        update_repo repo, file, data
       rescue Rugged::OSError, Rugged::RepositoryError => open_error
         begin
           Rugged::Repository.init_at repo, :bare
@@ -178,27 +178,24 @@ module Oxidized
       end
     end
 
-    def update_repo repo, file, data, msg, user, email
+    def update_repo repo, file, data
+      oid_old = repo.blob_at(repo.head.target_id, file) rescue nil
+      return false if oid_old and oid_old.content == data
+
       oid = repo.write data, :blob
       index = repo.index
-      index.read_tree repo.head.target.tree unless repo.empty?
+      index.add path: file, oid: oid, mode: 0100644
 
-      tree_old = index.write_tree repo
-      index.add :path => file, :oid => oid, :mode => 0100644
-      tree_new = index.write_tree repo
+      repo.config['user.name']  = @user
+      repo.config['user.email'] = @email
+      @commitref = Rugged::Commit.create(repo,
+                                         tree:       index.write_tree(repo),
+                                         message:    @msg,
+                                         parents:    repo.empty? ? [] : [repo.head.target].compact,
+                                         update_ref: 'HEAD')
 
-      if tree_old != tree_new
-        repo.config['user.name']  = user
-        repo.config['user.email'] = email
-        @commitref = Rugged::Commit.create(repo,
-                                           :tree       => index.write_tree(repo),
-                                           :message    => msg,
-                                           :parents    => repo.empty? ? [] : [repo.head.target].compact,
-                                           :update_ref => 'HEAD',)
-
-        index.write
-        true
-      end
+      index.write
+      true
     end
   end
 end