+++ lib/oxidized/string.rb
@@ -4,18 +4,18 @@ module Oxidized
     attr_accessor :type, :cmd, :name
 
     # @return [Oxidized::String] copy of self with last line removed
-    def cut_tail
-      Oxidized::String.new each_line.to_a[0..-2].join
+    def cut_tail lines = 1
+      Oxidized::String.new each_line.to_a[0..-1 - lines].join
     end
 
     # @return [Oxidized::String] copy of self with first line removed
-    def cut_head
-      Oxidized::String.new each_line.to_a[1..-1].join
+    def cut_head lines = 1
+      Oxidized::String.new each_line.to_a[lines..-1].join
     end
 
     # @return [Oxidized::String] copy of self with first and last lines removed
-    def cut_both
-      Oxidized::String.new each_line.to_a[1..-2].join
+    def cut_both head = 1, tail = 1
+      Oxidized::String.new each_line.to_a[head..-1 - tail].join
     end
 
     # sets @cmd and @name unless @name is already set